Adding 14 dots and 7 slashes
Sometimes we should fix copy-paste issues in our code base.

Because we can't just sit and wait for two or three hundred eternities
when a very promising AI-based coding agent does this for us.

If we have missing dots and slashes they should appear somehow in the
right places (just where it's neeeded).  Not very exciting, not very
creative job, but someone had to add them.  Usually we do it by
ourselves, by our hands, by typing missing dots and slashes on
keyboard (or copy-paste).