I don't care if tests fail in test profiles.
Sometimes we don't care of tests.

Because nobody will fix this.  Somebody should... but nobody would.
If it builds, runs, works and doing something it's already a big luck
and a great success.

Shurely sometimes it leads to SW/HW misbehaviour.  But it's rare
cases.  Hopefully...  Who knows?  Certainly nobody cares.  Tests fail
permanently.  That's life.  There is no fix for this or time to fix
this right now.